public/data/repos/educational_tutor_mcp/content/requirements.txt
openai
gradio
httpx
requests
python-dotenv
Pillow